+++ Assets/Hotfix/Runtime/ECS/Utility/WorldUtility.cs
@@ -2,7 +2,7 @@ namespace LccHotfix
 {
     public static class WorldUtility
     {
-        public static LogicEntity GetEntity(int id)
+        public static LogicEntity GetEntity(long id)
         {
             return WorldManager.Instance.GetWorld().GetEntityWithComID(id);
         }
